requirements.txt
# requirements.txt

# Core ML & CV Libraries
ultralytics
opencv-python-headless
numpy
filterpy
scikit-image

# FastAPI Backend
fastapi
uvicorn
python-multipart
Pillow

# Streamlit Frontend
streamlit
requests